docs(repo): update contributing flle to mention updating changelog

+++ src/main/java/automenta/vivisect/TreeMLData.java
@@ -208,7 +208,7 @@ public class TreeMLData implements MLData {
 
     @Override
     public int size() {
-        return (int) (getEnd() - getStart());
+        return getEnd() - getStart();
     }
